Temporary hack to fix Jekyll rendering
For some still unclear reason Jekyll is rendering assignment files
differently depending on their names. Nested list rendering works
fine for names starting with programming- and stuff- but not for
files starting with advanced- or databases- (or a number of other
naming conventions.

This needs to be solved more generally, but class starts in two
hours so this will have to do.