Fix display of exercises with lists in assignments
Jekyll currently processes pages in alphabetical order. This results
in cases where, when the name of an exercise included in an assignment
comes after the name of the assignment alphabetically, the
assignment page is processed before the exercise and the list in the
exercise is not properly converted to a list.

This implements a temporary hack to address this problem by renaming
all of the exercises to start with captical letters, and therefore
come before the names of the assignments (which are all lower case)
alphabetically.