Use new solution to Jekyll processing order issue
This reverts 1b1f509, which had temporarily changed the name of
one of the assignments to address the fact that Jekyll processes
pages alphabetically and this was causing issues with display of
lists on assignment pages.

This has now been fixed in a more general (but no less hacky) way
by naming exercise files starting with captical letters. See 79c5bbc.