Add instructor participation to Version Control 6
Version control exercises 6-9 are designed to involve an instructor who
is adding files to the repository and changing them to introduce
conflicts. This note informs readers of this and points them to the
relevant files in the repository so that instructors can use them and
self-lead students have access to them.