Integrate instruction on importing, types, and primary keys
Instead of talking and then showing later, this change integrates talking
about key aspects of databases with the process of creating the database in
the first place. Students are now actively using the tool much easier and it's
not necessary to show them a database separate from the one they are creating.