Update Loops-stringr-R to use loops more appropriately
The older exercise and solution weren't actually using the loop since data frames
in R loop over columns not rows. This updates the problem to actually use the loop
and in the process introduces indexes and pre-allocation of memory. This is a fair
bit of material to it switches to fill-in-the-blank for the for loop portion of the
exercise.