Introduce indexing and pre-allocation as part of looping over data frames
Looping over data frames in R is awkward, but it's a natural thing for students
to want to do. This brings that front and center and usese it to introduce indexing
and pre-allocation. This change is accompanied by an update to Loops-stringr-R.R
that adds all of these concepts to that exercise.