Switch conditionals lecture to tree mass estimate example
The superhero example introduced this concept with printing inside the
conditionals, which is not what most people would do and has lead to confusion
among students in the past. This material also only really worked well for
folk into superheros. This replaces it with something that models more standard
practice and is more easily accessible.