Update solutions to match standard approach
The current approach to shortening data output is
to print the output as it would occur in RStudio and
then remove all but the first 25 rows of output and
replace them with ... while keeping the "omitted rows"
numbers as they are. This wasn't followed in the recent
updates to the Working with Data exercises, which this corrects.