Creat new Solving Bigger Problems material
This builds on the old Computational Projects material, but brings
in introductions to style and debugging. The goal is to teach the
students how to start grappling with bigger problems that require
integrating material to solve problems that are more complex and
that less directly mimic the materials demonstrated in class.