Add backup copy of bioclim data in case web source goes down
The primary data source for `raster::getData` is a UC server
run by a single group. Sometimes, like today, it goes down
when you are about to have students work with it.
This backup copy of the data can be unzipped in the students
home directory and everything will work as expected. This
is `/home/user` on Linux and OSX and `C:\Users\user\Documents`
on Windows.