Update Loops-size-estimates-by-name-loop-R.md
Change wording in sentence before bullet point list to be more clear (still not entirely happy with it, but I think it's better).

Reformat the bullet point list because there is a lot of isolated code blocks and symbols that makes it hard to read. Related: having a single `=` between `species` and a value in this context may confuse some people that are not yet fully comfortable with the distinction between `=` and `==`.