Improve loops introduction
Separates the print statement when introducing print in loops to immediately show
that multiple lines can be part of the loop and avoid having to introduce this
separately later. This also removed the introduction of a variable that was only
used once in the lecture material and had to be removed in the next step.