Provide more guidance towards easiest approach in size-estimates-by-name exercises
These exercises where already rewritten to try to guide students towards
using `mass = a * length ^ b` in each conditional, but many are still
trying to set `a` and `b` which leads to confusion about how to handle
`mass = NA`. This revision attempts to be really explicit about using
the equations.